+++ libs/fslock/lock_wasm.go
@@ -3,16 +3,17 @@
 package fslock
 
 import (
-	"errors"
 	"log"
 )
 
 func (l *Locker) lock() (err error) {
 	log.Println("TODO: lock() implement me")
-	return errors.New("TODO: lock() implement me")
+	//return errors.New("TODO: lock() implement me")
+	return nil
 }
 
 func (l *Locker) unlock() error {
 	log.Println("TODO: unlock() implement me")
-	return errors.New("TODO: unlock() implement me")
+	//return errors.New("TODO: unlock() implement me")
+	return nil
 }
