+++ nodebuilder/keyring_wasm.go
@@ -1,10 +1,8 @@
-//go:build aix || darwin || dragonfly || freebsd || linux || netbsd || openbsd || solaris || zos
+//go:build wasm
 
 package nodebuilder
 
 import (
-	kr "github.com/cosmos/cosmos-sdk/crypto/keyring"
-
 	apptypes "github.com/celestiaorg/celestia-app/x/blob/types"
 
 	"github.com/celestiaorg/celestia-node/libs/keystore"
@@ -17,7 +15,7 @@ const DefaultAccountName = "my_celes_key"
 // NOTE: we construct keyring signer before constructing node for easier UX
 // as having keyring-backend set to `file` prompts user for password.
 func KeyringSigner(accName, backend string, ks keystore.Keystore, net p2p.Network) (*apptypes.KeyringSigner, error) {
-	ring := ks.Keyring()
+	/* ring := ks.Keyring()
 	var info *kr.Record
 	// if custom accName provided, find key for that name
 	if accName != "" {
@@ -35,12 +33,12 @@ func KeyringSigner(accName, backend string, ks keystore.Keystore, net p2p.Networ
 			return nil, err
 		}
 		info = keyInfo
-	}
+	} */
 	// construct signer using the default key found / generated above
-	signer := apptypes.NewKeyringSigner(ring, info.Name, string(net))
-	signerInfo := signer.GetSignerInfo()
-	log.Infow("constructed keyring signer", "backend", backend, "path", ks.Path(),
-		"key name", signerInfo.Name, "chain-id", string(net))
+	/* 	signer := apptypes.NewKeyringSigner(ring, info.Name, string(net))
+	   	signerInfo := signer.GetSignerInfo()
+	   	log.Infow("constructed keyring signer", "backend", backend, "path", ks.Path(),
+	   		"key name", signerInfo.Name, "chain-id", string(net)) */
 
-	return signer, nil
+	return &apptypes.KeyringSigner{}, nil
 }
